[PERFECTIVE] Enhanced the management of local memory map address blocks in the memory designer. Changed address calculations from 32-bit unsigned int to quint64.
git-svn-id: svn://svn.code.sf.net/p/kactus2/code/trunk@596 f0aa6db5-7f46-43cc-a8a5-a31efdaafa67

+++ MemoryDesigner/AddressBlockItem.cpp
@@ -19,8 +19,8 @@
 //-----------------------------------------------------------------------------
 AddressBlockItem::AddressBlockItem(QSharedPointer<Component> component,
                                    QSharedPointer<MemoryMap> memoryMap,
-                                   QString const& name, unsigned int startAddress,
-                                   unsigned int range, QGraphicsItem* parent)
+                                   QString const& name, quint64 startAddress,
+                                   quint64 range, QGraphicsItem* parent)
     : AddressSectionItem(name, startAddress, range, parent),
       component_(component),
       memoryMap_(memoryMap)
